python485master/requirements.txt
pymodbus==3.12.1
pyserial==3.5
